refonte optimized2

+++ models/actions.py
@@ -1,8 +1,8 @@
 class porteFolio:
     def __init__(self, name, cost, profit): 
         self.name = name
-        self.cost = int(cost)
-        self.profit = int(profit)
+        self.cost = float(cost)
+        self.profit = float(profit)
         self.gain = self.cost * (self.profit / 100 )
         self.ratio = (self.cost * (self.profit / 100)) + self.profit
